refactor: moved home page content to pages dir
